Rearrange source to follow normal Go packaging
The package name is the same name as the repository,
unless the repository will have lots of different packages
in it. This job is small enough to comfortably fit into
one package. Execuatables can go into sub-directories,
especially when they are demos. They will not be built
by "go get", because it works on one package at a time,
and the demos are in package main.

Finally, make the import path fully qualified. This is
a pain with the github "commit then request pull" workflow,
because you'll need to make a second commit to change the
import path in the demo back to your repo. See the closed
pull requests on github.com/huin/mqtt for the more normal
case, where I sent bug fixes in without changing the import
path of anything.