Simplify individual, added a test case
This is what tests look like in Go. Note that the
test is inside the genomego package, so it is allowed
to muck around with the internals of the object (in this
case to get rid of randomness, to facilitate testing).

Go does not use this. Instead, the receiver of the method
is chosen like a parameter name, i.e. short, but clear enough
to communicate what's happening. A good choice is always
the first letter of the object, like func (c *Client)Connect().
But in this case, that would be "i", which is "reserved" for
indexes, so a better choice is "me", since an Individual
would normally call itself "me" anyway, if it were alive.

I changed from len(me.genome) to me.size because it is clearer
and less noisy to read. Also, it sets us up for a refactoring
where we decouple the layout of genome from the size (i.e.
a bitfield).