Change to multi-value return for Crossover
Returning an array of kids is pointless here, and makes
garbage, because the [2]*Individual needs to be allocated
and then filled, then it's pointer goes onto the stack. The
current runtime implements multiple returns by pushing
each of them onto the stack, so this implementation creates
no garbage either on exit from Crossover or when the return
values arrive back in the caller (since the caller probably
has registers free to receive them and then work with them,
at least on AMD64. x86 is always out of registers, so it's always
spilling to memory...)

Also, naughty me. Last checkin did not have "go fmt" run on it.
See the difference? :)