Introduce a population type
Use it to hold a lambda that generates random numbers.

Use this to make the system more testable, using random
number generators that always give 0 or 1. Because Mutate
needs random numbers, Individuals now need to remember what
Population they belong to. :(

Note in NewPopulation that for every Population, we create
a new *math/rand.Rand, then capture it in the new closure
assigned to the new population's Rand field. If the Rand
field is overwritten later, then generator becomes unreferenced
and will be GC'd.