Fixed ga-example, fixed formatting of Individual
Adding a String() method to a type lets you control
how fmt prints it. If the object does not implement
the Stringer interface, then fmt will print all the
fields, which isn't what we want.

See:
	http://golang.org/src/pkg/fmt/print.go?s=1975:2019#L51
	http://golang.org/src/pkg/fmt/print.go#L702

fmt.Println(*father) didn't work because it dereferenced
father from an *Interface to an Interface, which does not
implement Stringer.